requirements.txt
PyMuPDF
pypdfium2
pdf2image
Pillow
PyYAML
python-dotenv
